[fold] matches_base is public
