+++ example/router/src.hpp
@@ -15,9 +15,9 @@
 #endif
 
 // include in one and no more than one source file
-#include "impl/router.ipp"
 #include "detail/impl/router.ipp"
-#include "detail/impl/matches_storage.ipp"
+#include "impl/matches.ipp"
+#include "impl/router.ipp"
 
 #endif
